+++ test/i18next.init.spec.js
@@ -1,11 +1,11 @@
-// i18next, v1.6.1
+// i18next, v1.6.3
 // Copyright (c)2013 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
 //////////////////////
 // HINT
 //
-// you need to replace '_fetchOne' with 'fetchOne' to use this on server
+// you need to replace 'fetchOne' with 'fetchOne' to use this on server
 //
 
 
@@ -23,7 +23,6 @@ describe('i18next.init', function() {
       fallbackLng: 'dev',
       fallbackNS: [],
       fallbackToDefaultNS: false,
-      fallbackOnNull: true,
       load: 'all',
       preload: [],
       supportedLngs: [],
@@ -32,11 +31,13 @@ describe('i18next.init', function() {
       resGetPath: 'test/locales/__lng__/__ns__.json',
       resSetPath: 'test/locales/__lng__/new.__ns__.json',
       saveMissing: false,
+      sendMissingTo: 'fallback',
       resStore: false,
       returnObjectTrees: false,
       interpolationPrefix: '__',
       interpolationSuffix: '__',
       postProcess: '',
+      parseMissingKey: '',
       debug: false
     };
 
@@ -308,7 +309,8 @@ describe('i18next.init', function() {
     
         describe('and fallbacking to multiple set namespace', function() {
           var resStore = {
-            dev: { 
+            dev: {
+              'ns.common': {},
               'ns.special': { 'simple_dev': 'ok_from_dev' },
               'ns.fallback1': { 
                 'simple_fallback': 'ok_from_fallback1',
@@ -337,6 +339,37 @@ describe('i18next.init', function() {
             expect(i18n.t('ns.common:simple_fallback')).to.be('ok_from_fallback1'); /* first wins */
             expect(i18n.t('ns.common:simple_fallback1')).to.be('ok_from_fallback1');
             expect(i18n.t('ns.common:simple_fallback2')).to.be('ok_from_fallback2');
+          });
+    
+          describe('and post missing', function() {
+    
+            var spy; 
+    
+            beforeEach(function(done) {
+              spy = sinon.spy(i18n.sync, 'postMissing');
+              i18n.init(i18n.functions.extend(opts, { 
+                fallbackNS: ['ns.fallback1', 'ns.fallback2'], 
+                resStore: resStore,
+                saveMissing: true,
+                ns: { namespaces: ['ns.common', 'ns.special', 'ns.fallback'], defaultNs: 'ns.special'} } ),
+                function(t) { 
+                  t('ns.common:notExisting');
+                  done(); 
+                });
+            });
+    
+            afterEach(function() {
+              spy.restore();
+            });
+    
+            it('it should post only to origin namespace', function() {
+              expect(spy.callCount).to.be(1);
+              expect(spy.args[0][0]).to.be('en-US');
+              expect(spy.args[0][1]).to.be('ns.common');
+              expect(spy.args[0][2]).to.be('notExisting');
+              expect(spy.args[0][3]).to.be('ns.common:notExisting');
+            });
+    
           });
     
         });