+++ releasenotes.md
@@ -1,7 +1,5 @@
-### next
-- fixes issue with saveMissing key/value - it should not throw error on unknown language namespace combination
-
 ### v1.6.8
+- fixes issue with saveMissing key/value - it should not throw error on unknown language namespace combination
 - fixes issue with fallbackLng = false in combination with sendMissingTo = fallback
 - updated client